refactor: put the header and productdetails in productdetails page
