+++ client/src/App.jsx
@@ -4,10 +4,9 @@ import './App.css';
 import SignUp from './Components/SignUp';
 import { Home } from './Components';
 import Login from './Components/Login';
-import Header from './Components/header/Header';
 // import ProtectedRoute from './Components/ProtectedRoute';
 import { AuthProvider } from './Hooks/useAuth';
-import ProductDetails from './Components/productdetails/ProductDetails';
+import ProductDetailsPage from './Pages';
 
 function App() {
   return (
@@ -15,7 +14,6 @@ function App() {
       {/* <Link to="/home">Home</Link>
       <Link to="/profile">Profile</Link> */}
       <AuthProvider>
-        <Header />
         <Routes>
           <Route path="/home" element={<Home />} />
           <Route path="/login" element={<Login />} />
@@ -29,7 +27,7 @@ function App() {
           /> */}
           {/* example how to use protected route */}
           <Route path="/signup" element={<SignUp />} />
-          <Route path="/product/:id" element={<ProductDetails />} />
+          <Route path="/product/:id" element={<ProductDetailsPage />} />
         </Routes>
       </AuthProvider>
     </div>
