+++ client/src/Pages/index.js
@@ -1,0 +1,3 @@
+import ProductDetailsPage from './ProductDetailsPage';
+
+export default ProductDetailsPage;
